# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API: Clients, Products and Orders",
        "description": "API: Client, Product and Orders",
        "contact": {
            "name": "Fernando Marcelino",
            "url": "https://github.com/fernandoomarcelino",
            "email": "ffernandomarcelino@gmail.com"
        },
        "license": {
            "name": "Private",
            "url": "https://github.com/fernandoomarcelino"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "L5 Swagger OpenApi dynamic host server"
        }
    ],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show all",
                "description": "Show all",
                "operationId": "ProductIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create",
                "description": "Create",
                "operationId": "ProductStore",
                "requestBody": {
                    "request": "ProductStoreRequest",
                    "description": "ProductStoreRequest",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Missing parameters (validation)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show detail",
                "description": "Show detail",
                "operationId": "ProductShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update",
                "description": "Update",
                "operationId": "ProductUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "ProductUpdateRequest",
                    "description": "ProductUpdateRequest",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Missing parameters (validation)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Remove",
                "description": "Remove",
                "operationId": "ProductDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProductStoreRequest": {
                "title": "ProductStoreRequest",
                "description": "The request to create product",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string"
                    },
                    "price": {
                        "title": "price",
                        "description": "price",
                        "type": "number",
                        "format": "number"
                    }
                },
                "type": "object"
            },
            "ProductUpdateRequest": {
                "title": "ProductUpdateRequest",
                "description": "The request to update product",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string"
                    },
                    "price": {
                        "title": "price",
                        "description": "price",
                        "type": "number",
                        "format": "number"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Validation Error response",
                "description": "The validation error response",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "error message",
                        "format": "string"
                    },
                    "errors": {
                        "title": "errors",
                        "description": "validation errors in the format: Property name: validation error",
                        "format": "array",
                        "items": {
                            "properties": {
                                "field": {
                                    "title": "field",
                                    "format": "string"
                                },
                                "value": {
                                    "title": "value",
                                    "format": "string"
                                }
                            },
                            "type": "object",
                            "format": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
    "security": [
        []
    ]
}